requirements.txt
Flask==2.2.2
Flask_Cors==4.0.0
numpy==1.24.1
pandas==1.5.3
PyJWT==2.8.0
pyotp==2.8.0
qrcode==7.4.2
Requests==2.31.0
scikit_learn==1.3.0
yfinance==0.2.38
